fonts: rename font loader to locator
